requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
google-generativeai==0.8.3
gunicorn==23.0.0
gevent==24.2.1
gevent-websocket==0.10.1
